pyproject.toml
[project]
name = "apc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adafruit-ads1x15>=1.0.2",
    "gpiozero>=2.0.1",
    "numpy>=2.3.1",
    "posix-ipc>=1.2.0",
    "rpi-gpio>=0.7.1",
    "simple-pid>=2.0.1",
]
